refactor: rename and replace profile components
