--- src/components/Profile/GreetingTitle/index.ts
@@ -1,2 +1,0 @@
-import GreetingTitle from './GreetingTitle';
-export { GreetingTitle };
