--- src/components/Profile/Tabs/FirstTab/UserDateOfBirth/UserDateOfBirth.module.scss
@@ -1,15 +1,0 @@
-@import '../../../../../styles/common';
-
-.root {
-  margin-bottom: 4rem;
-
-  &__container {
-    display: flex;
-    gap: 2.5rem;
-  }
-
-  &__error {
-    @include input-error()
-  }
-
-}
