--- src/components/Profile/Tabs/FirstTab/UserDateOfBirth/index.ts
@@ -1,2 +1,0 @@
-import UserDateOfBirth from './UserDateOfBirth';
-export { UserDateOfBirth };
