--- src/components/Profile/Tabs/FirstTab/index.ts
@@ -1,2 +1,0 @@
-import FirstTab from './FirstTab';
-export { FirstTab };
