--- src/components/Profile/Tabs/SecondTab/CustomerAdress/EditCard/editCard.module.scss
@@ -1,33 +1,0 @@
-@import '../../../../../../styles/common';
-
-.root {
-
-  &__title {
-    margin-bottom: 6rem;
-    padding-top: 2rem;
-    font-size: 4rem;
-    line-height: 4.4rem;
-  }
-
-  &__formContainer h3 {
-    display: none;
-  }
-
-  &__addressForm {
-    &> :last-child {
-      opacity: 0;
-      user-select: none;
-      pointer-events: none;
-    }
-  }
-
-  &__btn {
-    width: 100%;
-    height: 7rem;
-  }
-
-}
-
-.errorResponse {
-  @include responseMessage($light-primary, $error-color);
-}
