--- src/components/Profile/Tabs/SecondTab/CustomerAdress/EditCard/index.ts
@@ -1,2 +1,0 @@
-import EditCard from './EditCard';
-export { EditCard };
