--- src/components/Profile/Tabs/SecondTab/CustomerAdress/index.ts
@@ -1,2 +1,0 @@
-import CustomerAddress from './CustomerAddress';
-export { CustomerAddress };
