--- src/components/Profile/Tabs/SecondTab/Radiobtn/index.ts
@@ -1,2 +1,0 @@
-import Radiobtn from './Radiobtn';
-export { Radiobtn };
