--- src/components/Profile/Tabs/SecondTab/index.ts
@@ -1,2 +1,0 @@
-import SecondTab from './SecondTab';
-export { SecondTab };
