--- src/components/Profile/Tabs/SecondTab/secondTab.module.scss
@@ -1,77 +1,0 @@
-@import '../../../../styles/common';
-
-.root {
-  &__formContainer {
-    & h3 {
-      display: none;
-    }
-
-    &_form {
-      margin-bottom: 4rem;
-    }
-  }
-
-  &__btn {
-    width: 100%;
-    height: 7rem;
-    margin-bottom: 5rem;
-
-    &:disabled {
-      opacity: .5;
-    }
-  }
-
-  &__closeBtn {
-    @include close-btn();
-  }
-
-  &__addAddressBtn {
-    @include btn-primary();
-    margin-bottom: 4rem;
-    width: 100%;
-    height: 7rem;
-  }
-}
-
-.toggleShipping {
-  display: flex;
-  margin-bottom: 2rem;
-  border-bottom: 1px solid $dark;
-  text-transform: uppercase;
-
-  &__radio {
-    position: absolute;
-    width: 100%;
-    top: 0;
-    height: 100%;
-    opacity: 0;
-    left: 0;
-    cursor: pointer;
-  }
-
-  &__btn {
-    position: relative;
-    padding: 1.2rem 2rem;
-    line-height: 1.6rem;
-    font-size: 1.4rem;
-    font-weight: 600;
-
-    &_active {
-      border: 1px solid $dark;
-      border-bottom: none;
-
-      &::after {
-        content: '';
-        width: 100%;
-        display: block;
-        position: absolute;
-        border-bottom: 1px solid white;
-        z-index: 1;
-        height: 1px;
-        left: 0;
-        bottom: -1px;
-      }
-    }
-  }
-
-}
