--- src/components/Profile/Tabs/ThirdTab/index.ts
@@ -1,2 +1,0 @@
-import ThirdTab from './ThirdTab';
-export { ThirdTab };
