--- src/components/Profile/Tabs/index.ts
@@ -1,2 +1,0 @@
-import Tabs from './Tabs';
-export { Tabs };
