--- src/components/Profile/Tabs/tabs.module.scss
@@ -1,54 +1,0 @@
-@import '../../../styles/common';
-
-.root {
-  &__title {
-    margin-bottom: 6rem;
-    padding-top: 2rem;
-    font-size: 4rem;
-    line-height: 4.4rem;
-  }
-}
-
-.tabsList {
-  display: flex;
-  margin-bottom: 4rem;
-  border-bottom: 1px solid $dark;
-
-  &_hidden {
-    display: none;
-  }
-
-  &__tab {
-    position: relative;
-    padding: 1.2rem 2rem;
-    line-height: 120%;
-    font-size: 1.4rem;
-    font-weight: 600;
-    text-align: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-
-    @media(max-width: 800px) {
-      padding: 1.2rem 1.5rem;
-    }
-
-    &_active {
-      border: 1px solid $dark;
-      border-bottom: none;
-
-      &::after {
-        content: '';
-        width: 100%;
-        display: block;
-        position: absolute;
-        border-bottom: 2px solid white;
-        z-index: 1;
-        height: 1px;
-        left: 0;
-        bottom: -1px;
-      }
-    }
-  }
-}
