+++ src/components/ProfileTabs/AddressTab/AddressItem/addressItem.module.scss
@@ -1,0 +1,76 @@
+@import '../../../../styles/common/index';
+
+.item {
+  display: flex;
+  padding: 2rem 0;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  border-bottom: 0.1rem solid $dark;
+
+  &:first-child {
+    border-top: 0.1rem solid $dark;
+  }
+
+  &:last-child {
+    margin-bottom: 6rem;
+  }
+
+  &__btns {
+    display: flex;
+    flex-direction: column;
+  }
+
+  &__data {
+    @include font(regular, p1, $dark);
+    line-height: 22px;
+    margin-bottom: 2rem;
+
+    @media (max-width: 480px) {
+      font-size: 1.8rem;
+    }
+  }
+}
+
+.defaultSettContainer {
+  display: flex;
+  gap: 1rem;
+
+  flex-basis: 100%;
+  justify-content: start;
+
+  &__address {
+    padding: 0.2rem 0.7rem 0.3rem;
+    border-radius: 0.5rem;
+    font-size: 1.4rem;
+    background-color: rgba($dark, 0.08);
+    transition: opacity $transition-set;
+
+    &_active {
+      background-color: rgba($green, 0.3);
+      color: $dark-green;
+    }
+
+    &:hover {
+      opacity: 0.9;
+    }
+  }
+}
+
+.configureBtns {
+  &__btn {
+    @include font(regular, p1, $dark);
+    text-align: right;
+    line-height: 22px;
+    opacity: 0.3;
+    transition: opacity $transition-set;
+
+    &:hover {
+      color: $primary;
+      opacity: 1;
+    }
+
+    @media (max-width: 480px) {
+      font-size: 1.8rem;
+    }
+  }
+}
