+++ src/components/ProfileTabs/AddressTab/AddressItem/index.ts
@@ -1,0 +1,2 @@
+import AddressItem from './AddressItem.tsx';
+export { AddressItem };
