+++ src/components/ProfileTabs/AddressTab/AddressModal/index.ts
@@ -1,0 +1,2 @@
+import AddressModal from './AddressModal.tsx';
+export { AddressModal };
