+++ src/components/ProfileTabs/AddressTab/RadioButton/index.ts
@@ -1,0 +1,2 @@
+import RadioButton from './RadioButton.tsx';
+export { RadioButton };
