+++ src/components/ProfileTabs/AddressTab/index.ts
@@ -1,0 +1,2 @@
+import AddressTab from './AddressTab.tsx';
+export { AddressTab };
