+++ src/components/ProfileTabs/PasswordTab/index.ts
@@ -1,0 +1,2 @@
+import PasswordTab from './PasswordTab.tsx';
+export { PasswordTab };
