+++ src/components/ProfileTabs/PasswordTab/passwordTab.module.scss
@@ -1,0 +1,44 @@
+@import '../../../styles/common/index';
+
+.root {
+  &__editBtn {
+    @include btn-primary();
+    width: 100%;
+    height: 7rem;
+  }
+
+  &__label {
+    @include font(semibold, p1, $dark);
+    margin-bottom: 0.5rem;
+    line-height: 2.2rem;
+  }
+
+  &__input {
+    margin-bottom: 4rem;
+    background-color: inherit;
+    border: none;
+  }
+}
+
+.rootEdit {
+
+  &__input {
+    margin-bottom: 1rem;
+  }
+
+  &__closeBtn {
+    @include close-btn()
+  }
+}
+
+.form {
+
+  &__button {
+    width: 100%;
+    height: 7rem;
+
+    &:disabled {
+      opacity: 0.5;
+    }
+  }
+}
