+++ src/components/ProfileTabs/UsedDataTab/DateFields/dateFields.module.scss
@@ -1,0 +1,15 @@
+@import '../../../../styles/common/index';
+
+.root {
+  margin-bottom: 4rem;
+
+  &__container {
+    display: flex;
+    gap: 2.5rem;
+  }
+
+  &__error {
+    @include input-error()
+  }
+
+}
