+++ src/components/ProfileTabs/UsedDataTab/index.ts
@@ -1,0 +1,2 @@
+import UsedDataTab from './UsedDataTab.tsx';
+export { UsedDataTab };
