+++ src/components/ProfileTabs/UsedDataTab/usedDataTab.module.scss
@@ -1,0 +1,92 @@
+@import '../../../styles/common/index';
+
+.form {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 4rem;
+
+  &__input {
+    margin-bottom: 1.5rem;
+
+    label {
+      span {
+        font-weight: 600;
+      }
+
+      input {
+        border: none;
+        padding: 0;
+        color: $dark;
+      }
+    }
+  }
+
+  &__email {
+    margin-bottom: 4.5rem;
+
+    @media (max-width: 480px) {
+      margin-bottom: 4rem;
+    }
+  }
+
+  &__dateOfBirth {
+    position: relative;
+    margin-bottom: 3rem;
+
+    &::before {
+      content: 'Date of Birth';
+      position: absolute;
+      top: -2.5rem;
+      @include font(semibold, p1, $dark);
+      line-height: 2.2rem;
+
+      @media (max-width: 480px) {
+        font-size: 18px;
+        top: -2.4rem;
+      }
+    }
+  }
+}
+
+.formEdit {
+  margin-bottom: 4rem;
+
+  &__email {
+    margin-bottom: 1rem;
+  }
+
+  &__input {
+    margin-bottom: 1.5rem;
+
+    label {
+      span {
+        font-weight: 600;
+      }
+
+      input {
+        border: none;
+        padding: 0.5rem 0;
+        color: $dark;
+        border-bottom: 1px solid $dark;
+      }
+    }
+  }
+
+  // &__input input {
+  //   border-bottom: 1px solid $dark;
+  // }
+
+  &__closeBtn {
+    @include close-btn();
+  }
+
+  &__btn {
+    @include btn-primary();
+    width: 100%;
+    height: 7rem;
+
+    &:disabled {
+      opacity: 0.5;
+    }
+  }
+}
