+++ src/components/ProfileTabs/index.ts
@@ -1,0 +1,2 @@
+import ProfileTabs from './ProfileTabs.tsx';
+export { ProfileTabs };
