+++ src/constant/aboutus.ts
@@ -34,7 +34,7 @@ export const studentDataTabs: IAboutDataTabs[] = [
         'Work with the server SDK (Profile)',
         'Store setting (User)',
         'Login form implementation',
-        'Profile page implementation incl. Tabs',
+        'Profile page implementation incl. ProfileTabs',
         'About page implementation',
         'Trello board setup',
         'Work with React Testing Library',
