+++ src/pages/Profile/GreetingTitle/index.ts
@@ -1,0 +1,2 @@
+import GreetingTitle from './GreetingTitle.tsx';
+export { GreetingTitle };
